requirements.txt
#Run command -> pip install -r requirements.txt

#Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
httpx==0.27.0


# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.11
asyncpg==0.29.0

#Configuration
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1

#Utilities
pytz==2024.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

#Logging
structlog==24.2.0

#Testing
pytest==8.2.2
pytest-asyncio==0.23.7


# Data Pipelines
numpy==2.2.6
pandas-ta              # ATR, VWAP, technical indicators
yfinance==1.2.0                # Free market data (XAUUSD, DXY, US10Y)
apscheduler==3.10.4             # Cron jobs for scheduled data fetching
